Moved the code that initializes and stops the volume service to a separate object
